[ros] image bundle code to image bundler header/cpp files
