+++ cost_map_ros/include/cost_map_ros/converter.hpp
@@ -43,36 +43,6 @@
 namespace cost_map {
 
 /*****************************************************************************
-** Image Bundles
-*****************************************************************************/
-
-/**
- * @brief Initialises a adds a single layer from a yaml/image resource pair.
- *
- * @warning this will change the geometry of the provided costmap and delete all layers!
- * @todo bool result/exception handling for when things go wrong
- *
- * @param[in] filename : yaml file
- * @param[out] cost_map :
- */
-void fromImageBundle(const std::string& filename, cost_map::CostMap& cost_map);
-/**
- * @brief Dump a cost map to an image bundle set of files.
- *
- * This creates the specified yaml file with image bundle meta information
- * and a set of png images alongside, one for each layer in the cost map.
- *
- * @todo bool result/exception handling for when things go wrong
- *
- * @param[in] filename : name of the yaml file to write
- * @param[in] cost_map : cost map to dump
- */
-void toImageBundle(const std::string& filename, const cost_map::CostMap& cost_map);
-
-// this might be an interesting api to have too
-// void addLayerFromImageFile()
-
-/*****************************************************************************
 ** CostMap and GridMap
 *****************************************************************************/
